+++ src/kaos/plugins/tools/__init__.py
@@ -1,0 +1,22 @@
+"""KAOS tools: capabilities agents can invoke (implementations of Tool)."""
+
+from __future__ import annotations
+
+from kaos.plugins.tools.dev_tools import (
+    DEFAULT_ALLOWED_COMMANDS,
+    ListDirTool,
+    ReadFileTool,
+    RunCommandTool,
+    SearchCodeTool,
+    default_dev_tools,
+)
+
+__all__ = [
+    "DEFAULT_ALLOWED_COMMANDS",
+    "ListDirTool",
+    "ReadFileTool",
+    "RunCommandTool",
+    "SearchCodeTool",
+    "default_dev_tools",
+]
+